refactor: adiciona tratamento usando try/except

+++ src/apps/base/signals.py
@@ -11,19 +11,24 @@ from apps.transaction.models import RecurringTransactionModel, TransactionModel,
 @receiver([post_save, post_delete], sender=TransactionModel)
 @receiver([post_save, post_delete], sender=TransferModel)
 def clear_dashboard_cache(sender, instance, **kwargs):
-    if sender == AccountModel:
-        user_id = instance.user.id
+    try:
+        if sender == AccountModel:
+            user_id = instance.user_id
 
-    elif sender == TransactionModel or sender == RecurringTransactionModel:
-        user_id = instance.account.user.id
+        elif sender == TransactionModel or sender == RecurringTransactionModel:
+            user_id = instance.account.user_id
 
-    elif sender == TransferModel:
-        user_id = instance.original_account.user.id
-
-    index_key = f'user_dashboard_index_{user_id}'
-    keys_to_delete = cache.get(index_key, set())
-
-    if keys_to_delete:
-        cache.delete_many(list(keys_to_delete))
+        elif sender == TransferModel:
+            user_id = instance.original_account.user_id
     
-    cache.delete(index_key)
+    except Exception as e:
+        return
+
+    if user_id:
+        index_key = f'user_dashboard_index_{user_id}'
+        keys_to_delete = cache.get(index_key, set())
+
+        if keys_to_delete:
+            cache.delete_many(list(keys_to_delete))
+        
+        cache.delete(index_key)
